requirements.txt
fastapi
uvicorn
easyocr
torch
numpy
pillow
opencv-python-headless
pdf2image